combine py3 funcutils tests
